Calculator project: sum , subtract, multiple, divide

+++ src/main/java/com/example/spring_web_api/request/CalculatorInPut.java
@@ -1,6 +1,6 @@
 package com.example.spring_web_api.request;
 
 public class CalculatorInPut {
-    public Integer num1;
-    public Integer num2;
+    public Double num1;
+    public Double num2;
 }
